requirements.txt
streamlit==1.30.0
tensorflow==2.16.2
numpy==1.26.3
Pillow==9.5.0
pandas==1.5.3
matplotlib==3.7.1
seaborn==0.11.2
scikit-image==0.20.0
